refactor(ces): move ces resource to services path
