+++ huaweicloud/provider.go
@@ -16,6 +16,7 @@ import (
 	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/services/cce"
 	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/services/cci"
 	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/services/cdm"
+	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/services/ces"
 	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/services/css"
 	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/services/dcs"
 	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/services/dds"
@@ -460,7 +461,7 @@ func Provider() *schema.Provider {
 			"huaweicloud_cdm_job":                         cdm.ResourceCdmJob(),
 			"huaweicloud_cdm_link":                        cdm.ResourceCdmLink(),
 			"huaweicloud_cdn_domain":                      resourceCdnDomainV1(),
-			"huaweicloud_ces_alarmrule":                   ResourceAlarmRule(),
+			"huaweicloud_ces_alarmrule":                   ces.ResourceAlarmRule(),
 			"huaweicloud_cloudtable_cluster":              resourceCloudtableClusterV2(),
 			"huaweicloud_compute_instance":                ResourceComputeInstanceV2(),
 			"huaweicloud_compute_interface_attach":        ResourceComputeInterfaceAttachV2(),
